Runwal 25HourLife Manpada Thane
Runwal Group presents Runwal 25HourLife as an icon of an exemplary lifestyle, yet another masterpiece that's the epitome of artistry. This iconic address will be transformed into a tailor-made destination for the new generation. Modern, cosmopolitan and heaped in culture, this ongoing residential project is exceptionally defined for its novel style of living and excellent connectivity.

Life at Runwal 25 Hour Life is never less than a luxury and as well ensures an energetic state of mind for all ages, with an array of lifestyle amenities chosen only for you.

This majestic premium project will be expanded across 22 acres of lush green land. The development encourages you and also gives you an experience of a worthy lifestyle. Bang on Ghodbunder Road in Manpada, Thane the estate is as elite as its name. With these lavish homes and amenities, Manpada, Thane is encircled by neighbourhoods like Mulund, Bhandup, Kharegaon, Balkum and Anand Nagar. There are many modes of transport in Manpada that embrace Uber, taxicabs, and buses and the nearest railway station is Thane.

This amazing project comprises of 2BHKs and 3BHK apartments in 50 high-rise towers. Runwal 25 Hour Life Apartment offers the chance of going beyond the normal and provides a standard-of-living you always wished for. This housing development will be the venue for countless happy memories and is a perfect address for a serene life. The floor plan of the project offers 2-bed apartments (550 - 600 carpet) starting from 1.10Cr + Taxes and 3-bed apartments (625 - 650 carpet) starting from 1.25 Cr + Taxes

These new Runwal Project In Manpada Thane quote the most effective spacious homes with ample ventilation through large windows and 24*7 surveillance in the common area assuring you're a safe stay in your haven. Besides this elegant amenities plays an important role in crafting the luxuriousness here. These facilities accommodate an indoor games club, a gymnasium for fitness freaks, a yoga & meditation corner, a swimming pool, a play space for youths, and tons many more. These amenities in the project are those that each homebuyer aspires for.

This Runwal property and the epicentre Manpada is a rising location with a residential project near the financial capital of India Thane. Being the most premium space in Manpada close to famous Hospitals like Titan Hospital, Metro Hospital and R-Mall, Viviana Mall. Its proximity is incredibly close to schools, colleges (Carmel High School, Laurel Public School) and planned corporate campuses.
To know more - https://bit.ly/3Dlyg2h